P1-ui-optimization/requirements.txt
gurobipy==12.0.1
numpy==2.2.2
pillow==11.1.0
